+++ test/libsolidity/ABIJson/multiple_methods.sol
@@ -1,6 +1,6 @@
 contract test {
-    function f(uint a) public returns (uint d) { return a * 7; }
-    function g(uint b) public returns (uint e) { return b * 8; }
+    function f(uint256 a) public returns (uint256 d) { return a * 7; }
+    function g(uint256 b) public returns (uint256 e) { return b * 8; }
 }
 // ----
 //     :test
