+++ test/libsolidity/ABIJson/structs_in_libraries.sol
@@ -1,7 +1,7 @@
 pragma abicoder               v2;
 library L {
-    struct S { uint a; T[] sub; bytes b; }
-    struct T { uint[2] x; }
+    struct S { uint256 a; T[] sub; bytes b; }
+    struct T { uint256[2] x; }
     function f(L.S storage s) public view {}
     function g(L.S memory s) public view {}
 }
