+++ test/libsolidity/semanticTests/array/function_array_cross_calls.sol
@@ -1,6 +1,6 @@
 contract D {
-    function f(function() external returns (function() external returns (uint))[] memory x)
-        public returns (function() external returns (uint)[3] memory r) {
+    function f(function() external returns (function() external returns (uint256))[] memory x)
+        public returns (function() external returns (uint256)[3] memory r) {
         r[0] = x[0]();
         r[1] = x[1]();
         r[2] = x[2]();
@@ -10,11 +10,11 @@ contract D {
 
 contract C {
     function test() public returns (uint256, uint256, uint256) {
-        function() external returns (function() external returns (uint))[] memory x =
-            new function() external returns (function() external returns (uint))[](10);
+        function() external returns (function() external returns (uint256))[] memory x =
+            new function() external returns (function() external returns (uint256))[](10);
         for (uint256 i = 0; i < x.length; i++) x[i] = this.h;
         x[0] = this.htwo;
-        function() external returns (uint)[3] memory y = (new D()).f(x);
+        function() external returns (uint256)[3] memory y = (new D()).f(x);
         return (y[0](), y[1](), y[2]());
     }
 
@@ -32,11 +32,11 @@ contract C {
 
     uint256 counter;
 
-    function h() public returns (function() external returns (uint)) {
+    function h() public returns (function() external returns (uint256)) {
         return counter++ == 0 ? this.f : this.g;
     }
 
-    function htwo() public returns (function() external returns (uint)) {
+    function htwo() public returns (function() external returns (uint256)) {
         return this.e;
     }
 }
