+++ test/libsolidity/semanticTests/constructor_ihneritance_init_order_2.sol
@@ -1,11 +1,11 @@
 contract A {
-    uint x = 42;
+    uint256 x = 42;
     function f() public returns(uint256) {
         return x;
     }
 }
 contract B is A {
-    uint public y = f();
+    uint256 public y = f();
 }
 // ====
 // compileViaYul: also
