+++ test/libsolidity/semanticTests/libraries/internal_library_function_bound_to_internal_function.sol
@@ -1,17 +1,17 @@
 library L {
-    function double(function(uint) internal pure returns (uint) f, uint x) internal pure returns (uint) {
+    function double(function(uint256) internal pure returns (uint256) f, uint256 x) internal pure returns (uint256) {
         return f(x) * 2;
     }
 }
 
 contract C {
-    using L for function(uint) internal pure returns (uint);
+    using L for function(uint256) internal pure returns (uint256);
 
-    function identity(uint x) internal pure returns (uint) {
+    function identity(uint256 x) internal pure returns (uint256) {
         return x;
     }
 
-    function test(uint value) public returns (uint) {
+    function test(uint256 value) public returns (uint256) {
         return identity.double(value);
     }
 }
