+++ test/libsolidity/semanticTests/viaYul/conversion/explicit_cast_local_assignment.sol
@@ -1,5 +1,5 @@
 contract C {
-    function f(uint a) public pure returns (uint8 x) {
+    function f(uint256 a) public pure returns (uint8 x) {
         uint8 b = uint8(a);
         x = b;
     }
