+++ test/libsolidity/semanticTests/viaYul/local_assignment.sol
@@ -1,6 +1,6 @@
 contract C {
-    function f(uint a) public pure returns (uint x) {
-        uint b = a;
+    function f(uint256 a) public pure returns (uint256 x) {
+        uint256 b = a;
         x = b;
     }
 }
