+++ test/libsolidity/smtCheckerTests/functions/internal_call_with_assertion_inheritance_1.sol
@@ -1,7 +1,7 @@
 pragma experimental SMTChecker;
 
 contract A {
-	uint x;
+	uint256 x;
 	function f() internal {
 		assert(x == 1);
 		--x;
