+++ test/libsolidity/smtCheckerTests/inheritance/state_variables.sol
@@ -1,13 +1,13 @@
 pragma experimental SMTChecker;
 
 contract Base {
-	uint x;
-	uint z;
-	uint private t;
+	uint256 x;
+	uint256 z;
+	uint256 private t;
 }
 
 contract C is Base {
-	function f(uint y) public {
+	function f(uint256 y) public {
 		require(x < 10);
 		require(y < 100);
 		z = x + y;
