+++ test/libsolidity/smtCheckerTests/inheritance/state_variables_3.sol
@@ -1,14 +1,14 @@
 pragma experimental SMTChecker;
 
 contract Base {
-	uint x;
-	uint private t;
+	uint256 x;
+	uint256 private t;
 }
 
 contract C is Base {
 
-	uint private z;
-	function f(uint y) public {
+	uint256 private z;
+	function f(uint256 y) public {
 		require(x < 10);
 		require(y < 100);
 		z = x + y;
