+++ test/libsolidity/smtCheckerTests/loops/while_1.sol
@@ -2,7 +2,7 @@ pragma experimental SMTChecker;
 
 contract C
 {
-	function f(uint x, bool b) public pure {
+	function f(uint256 x, bool b) public pure {
 		require(x < 100);
 		while (x < 10) {
 			if (b)
