+++ test/libsolidity/smtCheckerTests/modifiers/modifier_overflow.sol
@@ -2,7 +2,7 @@ pragma experimental SMTChecker;
 
 contract C
 {
-	uint x;
+	uint256 x;
 
 	modifier m {
 		require(x > 0);
