+++ test/libsolidity/smtCheckerTests/operators/mod_even.sol
@@ -2,9 +2,9 @@ pragma experimental SMTChecker;
 
 contract C
 {
-	function f(uint x) public pure {
+	function f(uint256 x) public pure {
 		require(x < 10000);
-		uint y = x * 2;
+		uint256 y = x * 2;
 		assert((y % 2) == 0);
 	}
 }
