+++ test/libsolidity/smtCheckerTests/typecast/same_size.sol
@@ -11,7 +11,7 @@ contract C {
 		assert(x == -56);
 		int256 y = int256(uint256(2**255 + 10));
 		assert(y == -(2**255) + 10);
-		int256 z = int256(uint(2**255 + 10));
+		int256 z = int256(uint256(2**255 + 10));
 		assert(z == -(2**255) + 10);
 		int256 t = int256(uint256(bytes32(uint256(200))));
 		assert(t == 200);
@@ -19,7 +19,7 @@ contract C {
 		assert(v == -(2**255) + 10);
 		int160 a = int160(uint160(address(type(uint160).max)));
 		assert(a == -1);
-		int160 b = int160(uint160(address(uint160(uint(2**159 + 10)))));
+		int160 b = int160(uint160(address(uint160(uint256(2**159 + 10)))));
 		assert(b == -(2**159) + 10);
 		D d;
 		int160 e = int160(uint160(address(d)));
@@ -50,7 +50,7 @@ contract C {
 		address e = address(d);
 		assert(uint160(e) == 0);
 		E f = E(1);
-		assert(uint(f) == 1);
+		assert(uint256(f) == 1);
 	}
 
 	function f3() public pure {
