+++ test/libsolidity/syntaxTests/multiSource/circular_import.sol
@@ -1,6 +1,6 @@
 ==== Source: s1.sol ====
 import {f as g} from "s2.sol";
-function f() pure returns (uint) { return 1; }
+function f() pure returns (uint256) { return 1; }
 ==== Source: s2.sol ====
 import {f as g} from "s1.sol";
-function f() pure returns (uint) { return 2; }
+function f() pure returns (uint256) { return 2; }
